+++ towerhopscotch/src/main/java/ch/zhaw/psit/towerhopscotch/models/entities/enemies/Slime.java
@@ -7,7 +7,7 @@ import ch.zhaw.psit.towerhopscotch.models.maps.Layer;
 import java.awt.*;
 
 public class Slime extends Enemy {
-    public Slime(Layer onLayer, float x, float y, Player player) {
+    public Slime(Layer onLayer, int x, int y, Player player) {
         super(onLayer, x, y, Enemy.DEFAULT_WIDTH, Enemy.DEFAULT_HEIGHT, 100, 1, 1.0f, player, 10);
     }
 
@@ -16,6 +16,6 @@ public class Slime extends Enemy {
     }
 
     public void render(Graphics g) {
-        renderEnemy(g,Assets.enemies.get("slime"));
+        renderEnemy(g, Assets.enemies.get("slime"));
     }
 }
